# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "title": "REST API",
    "description": "REST API with Express and Firestore",
    "version": "1.0.0"
  },
  "host": "restapi-6a51.onrender.com",
  "basePath": "/api",
  "schemes": [
    "https"
  ],
  "paths": {
    "/api/items/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/items": {
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Create an item",
        "description": "Create an item",
        "parameters": [
          {
            "name": "data",
            "in": "body",
            "description": "Data to create an item",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Item successfully created"
          },
          "400": {
            "description": "Bad request"
          }
        }
      },
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get all items entries",
        "description": "Get all items entries",
        "responses": {
          "200": {
            "description": "Items entries successfully obtained"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/items/{id}": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get an item entry",
        "description": "Get an item entry",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "Item id"
          }
        ],
        "responses": {
          "200": {
            "description": "Get an item by id"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Item not found"
          }
        }
      },
      "put": {
        "tags": [
          "Items"
        ],
        "summary": "",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          },
          {
            "name": "data",
            "in": "body",
            "description": "",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      },
      "delete": {
        "tags": [
          "Items"
        ],
        "summary": "",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string",
            "description": ""
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        }
      }
    }
  }
}